Adding polyfill because the solution is much lighter
